+++ packages/axios-extension/src/auth/uaa.ts
@@ -12,7 +12,7 @@ import { UAATimeoutError } from './error';
 import { redirectSuccessHtml } from './static';
 import { defaultTimeout } from './connection';
 
-export type RefreshTokenChanged = (refreshToken?: string) => void | Promise<void>;
+export type RefreshTokenChanged = (refreshToken?: string, accessToken?: string) => void | Promise<void>;
 
 /**
  * A class representing interactions with an SAP BTP UAA service
@@ -290,9 +290,9 @@ export class Uaa {
             newRefreshToken = response.data.refresh_token;
         }
 
-        if (newRefreshToken && refreshTokenChangedCb) {
+        if (refreshTokenChangedCb) {
             this.log.info('Sending notification that refresh token changed');
-            await refreshTokenChangedCb(newRefreshToken);
+            await refreshTokenChangedCb(newRefreshToken, response?.data?.access_token);
         }
 
         this.log.info('Got access token successfully');